pkg/proc: Minor reorg
